fix(group-tasks): sync after claiming/unclaiming
