+++ website/client/components/header/menu.vue
@@ -349,10 +349,11 @@ import creatorIntro from '../creatorIntro';
 import InboxModal from '../userMenu/inbox.vue';
 import notificationMenu from './notificationsDropdown';
 import profileModal from '../userMenu/profileModal';
+import reportFlagModal from '../chat/reportFlagModal';
 import sendGemsModal from 'client/components/payments/sendGemsModal';
+import sync from 'client/mixins/sync';
 import userDropdown from './userDropdown';
 
-import reportFlagModal from '../chat/reportFlagModal';
 
 export default {
   components: {
@@ -364,6 +365,7 @@ export default {
     sendGemsModal,
     userDropdown,
   },
+  mixins: [sync],
   data () {
     return {
       isUserDropdownOpen: false,
@@ -402,14 +404,6 @@ export default {
     },
     toggleUserDropdown () {
       this.isUserDropdownOpen = !this.isUserDropdownOpen;
-    },
-    async sync () {
-      this.$root.$emit('habitica::resync-requested');
-      await Promise.all([
-        this.$store.dispatch('user:fetch', {forceLoad: true}),
-        this.$store.dispatch('tasks:fetchUserTasks', {forceLoad: true}),
-      ]);
-      this.$root.$emit('habitica::resync-completed');
     },
     async getUserGroupPlans () {
       this.$store.state.groupPlans = await this.$store.dispatch('guilds:getGroupPlans');
