+++ website/client/components/tasks/approvalFooter.vue
@@ -35,8 +35,10 @@ div
 import findIndex from 'lodash/findIndex';
 import { mapState } from 'client/libs/store';
 import approvalModal from './approvalModal';
+import sync from 'client/mixins/sync';
 
 export default {
+  mixins: [sync],
   props: ['task', 'group'],
   components: {
     approvalModal,
@@ -91,11 +93,12 @@ export default {
         taskId = this.task.group.taskId;
       }
 
-      this.$store.dispatch('tasks:assignTask', {
+      await this.$store.dispatch('tasks:assignTask', {
         taskId,
         userId: this.user._id,
       });
       this.task.group.assignedUsers.push(this.user._id);
+      this.sync();
     },
     async unassign () {
       if (!confirm(this.$t('confirmUnClaim'))) return;
@@ -106,12 +109,14 @@ export default {
         taskId = this.task.group.taskId;
       }
 
-      this.$store.dispatch('tasks:unassignTask', {
+      await this.$store.dispatch('tasks:unassignTask', {
         taskId,
         userId: this.user._id,
       });
       let index = this.task.group.assignedUsers.indexOf(this.user._id);
       this.task.group.assignedUsers.splice(index, 1);
+
+      this.sync();
     },
     approve () {
       let userIdToApprove = this.task.group.assignedUsers[0];
